feat: add local resolver
