+++ Cargo.toml
@@ -6,9 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+bytes = "1.5"
 aws-sdk-s3 = "0.31.0"
 maxminddb = "0.23"
-bytes = "1.5"
-
 thiserror = "1.0"
-justerror = "1.1"
